Add consoles command

+++ database/migrations/2020_08_18_121401_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             
             $table->bigIncrements('id');
             $table->timestamp('created_on');
-            $table->unsignedBigInteger('external_id');
+            $table->integer('external_id');
             $table->string('describe')->length(1000)->nullable()->default(NULL);
             $table->string('name')->length(200);
             $table->float('price', 8, 2);
